Update lock timeout calculation on heartbeat
Use the reqTime to calculate the new lock timeout on hearbeat.
This is not perfect as cannot know the exact tick that the server
updates the database, but the assertion in the release lock
validation will be correct because we err on the side of under
calculating the timeout time.