Create callback request returns 200 (#406)
When a request to create a callback is made against a promise that
is already completed, the response will contain a 200 code along
with the promise information. The caller can use this information
to immediately continue. The 200 response code contrasts to a 201
that is returned in the case where the promise is pending and a
callback is created.