DST horizontal partitioning (#556)
This PR contains several related changes to be able to make the DST run faster and enable validations between promises and tasks.

Make DST do horizontal partitions based on the rootPromiseId
Add many validations between tasks and promises as well as adjusting some others
Make a complete promise operation also complete all related tasks including the current claimed task
General small changes to the way we print and report DST results to make it easier to debug.
Make trying to complete a task that is already completed return a StatusOk response
Create a single CreatePromiseAndTask command that is aware of not creating a task in case a promise was not being able to be created.