Use task as fencing token (#813)
This PR adds support for using a claimed task as a fencing token for create promise, complete promise and crate callback, operations. This means that if the task is not currently claimed the operation is not performed and an error is sent back to the user. The task information must is sent in the header of the requests as `task-id` and `task-counter`